requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.1
pydantic-settings==2.5.2
pymongo==4.10.1
mongomock==4.1.2
pinecone-client==5.0.1
openai==1.47.0
PyYAML==6.0.2
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.1.1
httpx==0.27.2
pytest==8.3.3
mongomock==4.1.2
